Move yamaha_musiccast coordinator to separate module
