+++ homeassistant/components/yamaha_musiccast/select.py
@@ -9,8 +9,8 @@ from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from . import DOMAIN, MusicCastDataUpdateCoordinator
-from .const import TRANSLATION_KEY_MAPPING
+from .const import DOMAIN, TRANSLATION_KEY_MAPPING
+from .coordinator import MusicCastDataUpdateCoordinator
 from .entity import MusicCastCapabilityEntity
