+++ homeassistant/components/yamaha_musiccast/switch.py
@@ -9,7 +9,8 @@ from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from . import DOMAIN, MusicCastDataUpdateCoordinator
+from .const import DOMAIN
+from .coordinator import MusicCastDataUpdateCoordinator
 from .entity import MusicCastCapabilityEntity
